refactor(segmented-control): merge into single file
