+++ packages/bezier-react/src/components/Forms/SegmentedControl/SegmentedControl.tsx
@@ -1,5 +1,6 @@
 import React, {
   forwardRef,
+  useEffect,
   useMemo,
   useState,
 } from 'react'
@@ -8,13 +9,22 @@ import * as RadioGroupPrimitive from '@radix-ui/react-radio-group'
 import * as TabsPrimitive from '@radix-ui/react-tabs'
 import classNames from 'classnames'
 
-import { cssDimension } from '~/src/utils/style'
+import { ariaAttr } from '~/src/utils/dom'
+import { createContext } from '~/src/utils/react'
+import {
+  cssDimension,
+  px,
+} from '~/src/utils/style'
 
-import { Divider } from '~/src/components/Divider'
+import {
+  DIVIDER_THICKNESS,
+  Divider,
+} from '~/src/components/Divider'
 import useFormFieldProps from '~/src/components/Forms/useFormFieldProps'
 
 import {
   type SegmentedControlItemListProps,
+  type SegmentedControlItemProps,
   type SegmentedControlProps,
   type SegmentedControlRadioGroupProps,
   SegmentedControlSize,
@@ -23,16 +33,60 @@ import {
   type SegmentedControlTabsProps,
   type SegmentedControlType,
 } from './SegmentedControl.types'
-import {
-  SegmentedControlContextProvider,
-  SegmentedControlItemContextProvider,
-  SegmentedControlItemListContextProvider,
-  type SegmentedControlItemListContextValue,
-  useSegmentedControlContext,
-} from './SegmentedControlContext'
-import { SegmentedControlIndicator } from './SegmentedControlIndicator'
 
 import * as Styled from './SegmentedControl.styled'
+
+type SegmentedControlContextValue = Required<Pick<SegmentedControlProps<SegmentedControlType, string>, 'type' | 'size' | 'width'>>
+
+const [
+  SegmentedControlContextProvider,
+  useSegmentedControlContext,
+] = createContext<SegmentedControlContextValue | null>(null, 'SegmentedControl')
+
+type SegmentedControlItemListContextValue = {
+  itemCount: number
+  selectedItemIndex: number | null
+  setSelectedItemIndex: (index: number | null) => void
+}
+
+const [
+  SegmentedControlItemContextProvider,
+  useSegmentedControlItemContext,
+] = createContext<number | null>(null, 'SegmentedControlItem')
+
+const [
+  SegmentedControlItemListContextProvider,
+  useSegmentedControlItemListContext,
+] = createContext<SegmentedControlItemListContextValue | null>(null, 'SegmentedControlItemList')
+
+export const SEGMENTED_CONTROL_INDICATOR_TEST_ID = 'bezier-react-segmented-control-indicator'
+
+function SegmentedControlIndicator() {
+  const { selectedItemIndex, itemCount } = useSegmentedControlItemListContext('SegmentedControlIndicator')
+  const { size } = useSegmentedControlContext('SegmentedControlIndicator')
+
+  if (selectedItemIndex === null) { return null }
+
+  const containerPadding = Styled.paddingBySegmentedControlSize[size]
+  const containerHeight = Styled.heightBySegmentedControlSize[size]
+
+  const dividerTotalWidth = `${(itemCount - 1) * DIVIDER_THICKNESS}px`
+  const containerHorizontalPadding = `${2 * containerPadding}px`
+
+  const style = {
+    '--b-segmented-control-indicator-translateX': `calc(${selectedItemIndex * 100}% + ${selectedItemIndex * DIVIDER_THICKNESS}px)`,
+    '--b-segmented-control-indicator-width': `calc((100% - ${dividerTotalWidth} - ${containerHorizontalPadding}) / ${itemCount})`,
+    '--b-segmented-control-indicator-height': px(containerHeight - (2 * containerPadding)),
+    '--b-segmented-control-indicator-left': px(containerPadding),
+  } as React.CSSProperties
+
+  return (
+    <Styled.Indicator
+      data-testid={SEGMENTED_CONTROL_INDICATOR_TEST_ID}
+      style={style}
+    />
+  )
+}
 
 function SegmentedControlItemListImpl<
   Type extends SegmentedControlType,
@@ -223,4 +277,94 @@ export const SegmentedControl = forwardRef(SegmentedControlImpl) as <
   Value extends string,
 >(
   props: SegmentedControlProps<Type, Value> & { ref?: React.ForwardedRef<HTMLDivElement> }
+) => JSX.Element
+
+/**
+ * NOTE: (@ed) A property injected at runtime by the radix-ui lib.
+ */
+type ItemProps<Type extends SegmentedControlType> = (
+  Type extends 'radiogroup'
+    ? { 'data-state'?: 'unchecked' | 'checked' }
+    : { 'data-state'?: 'inactive' | 'active' }
+)
+& React.HTMLAttributes<HTMLButtonElement>
+& Partial<SegmentedControlItemProps<Type>>
+& Required<Pick<SegmentedControlProps<Type, string>, 'type'>>
+
+const Item = forwardRef<HTMLButtonElement, ItemProps<SegmentedControlType>>(function Item({
+  children,
+  type,
+  leftContent,
+  rightContent,
+  ...rest
+}, forwardedRef) {
+  const { setSelectedItemIndex } = useSegmentedControlItemListContext('SegmentedControlItem')
+  const index = useSegmentedControlItemContext('SegmentedControlItem')
+
+  const checked = type === 'radiogroup'
+    ? (rest as ItemProps<typeof type>)?.['data-state'] === 'checked'
+    : (rest as ItemProps<typeof type>)?.['data-state'] === 'active'
+
+  useEffect(function setSelectedItem() {
+    if (checked) {
+      setSelectedItemIndex(index)
+    }
+  }, [
+    checked,
+    index,
+    setSelectedItemIndex,
+  ])
+
+  return (
+    <Styled.Item
+      {...rest}
+      ref={forwardedRef}
+      type="button"
+      data-checked={ariaAttr(checked)}
+    >
+      <Styled.ItemContainer>
+        { leftContent }
+        <Styled.ItemLabel>
+          { children }
+        </Styled.ItemLabel>
+        { rightContent }
+      </Styled.ItemContainer>
+    </Styled.Item>
+  )
+})
+
+function SegmentedControlItemImpl<Value extends string>({
+  children,
+  ...rest
+}: SegmentedControlItemProps<Value>, forwardedRef: React.Ref<HTMLButtonElement>) {
+  const { type } = useSegmentedControlContext('SegmentedControlItem')
+
+  const SegmentedControlItem = type === 'radiogroup'
+    ? RadioGroupPrimitive.Item
+    : TabsPrimitive.Trigger
+
+  return (
+    <SegmentedControlItem
+      asChild
+      ref={forwardedRef}
+      {...rest}
+    >
+      <Item type={type}>
+        { children }
+      </Item>
+    </SegmentedControlItem>
+  )
+}
+
+/**
+ * `SegmentedControlItem` component is each item in `SegmentedControl`.
+ *
+ * If the type of `SegmentedControl` is `radiogroup`, this component acts as a radio item.
+ * In this case, it must be used as a child of `SegmentedControl`.
+ *
+ * If the type of `SegmentedControl` is `tabs`, this component acts as a tab item.
+ * In this case, it must be used as a child of `SegmentedControlTabList`.
+ */
+export const SegmentedControlItem = forwardRef(SegmentedControlItemImpl) as <Value extends string>(
+  props: SegmentedControlItemProps<Value> & { ref?: React.ForwardedRef<HTMLButtonElement> }
 ) => JSX.Element
